Add proof of concept for scraper
It's untested, and has poor error handling, but it does create the info
part of the openapi spec.

Some fields have been left as blank, and that's because I couldn't see
anything corresponding to them in the HTML document.

I'm not sure what I'll do about the error handling:
On the one hand it's not going to be part of a larger program, and the
input is completely static, so panicking may be ok.
On the other hand, using `unwrap` looks messy and is against my personal
rust style guide.