Extend schema parsing to do all sections
It currently only does string fields.
The next step will be to do the other types like internal references,
boolean, integer(int32), integer(int64), < _ > array, Map,
number(float), Object and enum (A, B, C).

It is likely that Map will lose information, as it simply isn't in the
HTML what the key and values are.

I've only included the sections which are purely strings as they're the
only working sections currently.