Avoid creating summaries for all path items
If a path item doesn't have a summary defined in the HTML, it shouldn't
end up with one in the OpenAPI.

In the HTML, usually the path is given in a pre tag and the summary in
an initial h4. If no summary is given, then the path is given in the
initial h4 and there is no pre tag.

So to find when we don't want a summary, we check to see if there is a
pre-tag or not.