Move most of the Selectors to be lazy static
I believe this is the rust way. Plus it looks neater.

There may be a very slight performance improvement when running the
tests. It isn't large enough to notice.

I think most of the run time is spent parsing the HTML which
unfortunately cannot be shared.

I should make the info module lazy static too, but that will require a slight
refactoring.