Hardcode out the /{any} path
It is invalid in the swagger editor and openapi generator because it
doesn't define all its paths.

I thought about faking its paths, but I don't think that would work
because this route is meant to cover sub paths.

I don't think this is an important route to have in the spec, because
it's designed for web browsers, not API users.