Remove cargo and npm as make-targets
Make doesn't work so well when something depends on a whole directory.

Also, I wasn't happy that to build the example app you had to build
the openapi definitions.

As such, I've taken the example app out of the makefile.

And whilst the transformer is in the makefile, it's done in a way where
cargo can handle its own dependencies.