Add tag support
I've made a tactical choice to tag everything with both the section from
the HTML documentation and the word "Default".

This is because (at least with the Typescript generator used in the
example), not tagging with Default breaks existing code. This is because
previously it would generate the operations within DefaultAPI, but
stopped doing that when other tags were present.

Requested by @yordis
https://github.com/ccouzens/keycloak-openapi/issues/2